Moving MSA alias from make_prg/from_msa/__init__.py to make_prg/__init__.py
