+++ make_prg/__init__.py
@@ -14,3 +14,6 @@ else:
 __version__ = metadata.version("make_prg")
 
 __all__ = ["from_msa", "subcommands"]
+
+from Bio.AlignIO import MultipleSeqAlignment
+MSA = MultipleSeqAlignment
