+++ make_prg/from_msa/__init__.py
@@ -1,6 +1,3 @@
 # ___Constants/Aliases___ #
-from Bio.AlignIO import MultipleSeqAlignment
-
-MSA = MultipleSeqAlignment
 NESTING_LVL = 5
 MIN_MATCH_LEN = 7
