+++ make_prg/from_msa/cluster_sequences.py
@@ -4,7 +4,7 @@ from itertools import starmap, repeat, chain
 from loguru import logger
 import numpy as np
 from sklearn.cluster import KMeans
-from make_prg.from_msa import MSA
+from make_prg import MSA
 from make_prg.utils.misc import flatten_list
 from make_prg.utils.seq_utils import ungap, Sequence, Sequences, SequenceExpander
 from dataclasses import dataclass
